fix: rename chat box
